requirements.txt
asn1crypto==0.23.0
astroid==1.5.3
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pylint==1.7.4
pyOpenSSL==17.5.0
python-binance==0.5.3
requests==2.18.4
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
wrapt==1.10.11
zope.interface==4.4.3
